Renamed `actions/epics.js` to `actions/index.js` and adjusted imports
