+++ app/src/components/MessageConversation/MessageConversation.jsx
@@ -8,7 +8,7 @@ import {
     addRecipients,
     cancelAttachment,
     downloadAttachment,
-} from '../../actions/epics'
+} from '../../actions'
 
 import { getInstance as getD2Instance } from 'd2/lib/d2'
 
@@ -90,13 +90,12 @@ class MessageConversation extends Component {
 
         const participants = messageConversation.userMessages
             .slice(0, maxParticipantsDisplay)
-            .map(
-                userMessage =>
-                    typeof this.state.currentUser === 'undefined' ||
-                    this.state.recipientsExpanded ||
-                    this.state.currentUser.id !== userMessage.user.id
-                        ? userMessage.user.displayName
-                        : i18n.t('me')
+            .map(userMessage =>
+                typeof this.state.currentUser === 'undefined' ||
+                this.state.recipientsExpanded ||
+                this.state.currentUser.id !== userMessage.user.id
+                    ? userMessage.user.displayName
+                    : i18n.t('me')
             )
 
         const userMessagesLength = messageConversation.userMessages.length
