+++ public/manifest.webapp
@@ -17,5 +17,5 @@
     "company": "DHIS2"
   },
   "description": "App for messaging",
-  "manifest_generated_at": "Thu Nov 15 2018 10:08:59 GMT+0100 (CET)"
+  "manifest_generated_at": "Thu Nov 15 2018 16:48:51 GMT+0100 (CET)"
 }
